+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## [1.8.0] - 2025-01-17
 ### Added
 * Added support for the Write API in the client library, the implementation
   can be found in the `DeepLClient` class. Please refer to the README for usage
@@ -135,7 +135,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial version.
 
 
-[Unreleased]: https://github.com/DeepLcom/deepl-java/compare/v1.7.0...HEAD
+[1.8.0]: https://github.com/DeepLcom/deepl-java/compare/v1.7.0...v1.8.0
 [1.7.0]: https://github.com/DeepLcom/deepl-java/compare/v1.6.0...v1.7.0
 [1.6.0]: https://github.com/DeepLcom/deepl-java/compare/v1.5.1...v1.6.0
 [1.5.1]: https://github.com/DeepLcom/deepl-java/compare/v1.5.0...v1.5.1